Cleanup and reorganize, fix race when switching images in OmeZarrImageViewer
Fix imports
