+++ packages/react/src/components/index.ts
@@ -3,11 +3,5 @@ export { default as App } from "./App";
 export { default as AppWithProviders } from "./AppWithProviders";
 export { IdetikCanvas } from "./IdetikCanvas";
 
-// Hooks
-export * from "./hooks";
-
-// Providers
 export * from "./providers";
-
-// Viewers
 export * from "./viewers";
