+++ packages/react/src/components/viewers/index.ts
@@ -1,2 +1,1 @@
-// OME-Zarr Image Viewer
 export * from "./OmeZarrImageViewer";
