+++ packages/react/src/index.ts
@@ -1,5 +1,5 @@
 // import css so it gets packaged with the library
 import "./input.css";
 
-// Re-export all components using barrel files
 export * from "./components";
+export * from "./hooks";
